Add tests for UnlockAccounts
In addition to the tests themselves, I have to add lines to .gitignore.
This is because the tests use the test directory as the "user's" home
directore and so the resultant dotfiles are created there. There is no
need for these to be added to git as they are incidental to the tests.